Refactor

+++ output/horse.rb
@@ -1,5 +1,5 @@
-require_relative '../config/settings.rb'
-require_relative '../http/client.rb'
+require_relative '../settings/settings.rb'
+require_relative '../client/http.rb'
 require 'fileutils'
 
 def output_horse(file_id)
@@ -12,6 +12,8 @@ def output_horse(file_id)
   [].tap do |horse_ids|
     horse_paths.each do |path|
       horse_id = path.match(/\/horse\/(\d+)/)[1]
+      next if File.exists?(File.join(output_dir, "#{horse_id}.html"))
+
       res = HTTPClient.new.get_horse(horse_id)
 
       File.open(File.join(output_dir, "#{horse_id}.html"), "w:utf-8") do |out|
